Add files via upload
Len Bielenberg was born in 1927 and was raised in Genesee, Idaho. After serving in World War 2, Len enrolled at the University of Idaho and graduated in 1950 with a business law degree and again in 1952 with a doctorate. Len has lived in Moscow, Idaho since 1954 and has had the unique experience of living in this town during the revolutionary decade that is the 1960s. Some of the major themes discussed in this interview are cultural changes, social changes, and political changes that occurred in the 1960s.